Rename AzureIot.h -> AzureIoTHub.h
