+++ examples/esp8266/simplesample_http/simplesample_http.c
@@ -12,7 +12,7 @@ That does not mean that HTTP only works with the _LL APIs.
 Simply changing the using the convenience layer (functions not having _LL)
 and removing calls to _DoWork will yield the same results. */
 
-#include "AzureIoT.h"
+#include "AzureIoTHub.h"
 
 static const char* connectionString = "HostName=[host].azure-devices.net;DeviceId=[device];SharedAccessKey=[key]";
