+++ examples/samd/remote_monitoring/remote_monitoring.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-#include "AzureIoT.h"
+#include "AzureIoTHub.h"
 #include "sdk/schemaserializer.h"
 #include "bme280.h"
